Add puzzle launch from standalone walkthrough viewer
