+++ src/routes/sudoku/[id]/+page.ts
@@ -1,10 +1,13 @@
 import trpc from '$lib/client/trpc';
 import { error } from '@sveltejs/kit';
 import type { PageLoad } from './$types';
+import { decompressFromBase64 } from '$utils/compressor';
+import type { SolutionStep } from '$models/Walkthrough';
 
-export const load: PageLoad = async ({ fetch, params }) => {
+export const load: PageLoad = async ({ fetch, params, url }) => {
   const trpcClient = trpc(fetch);
   const sudokuId = parseInt(params.id);
+  const dataParam = url.searchParams.get('data');
   const [sudoku, walkthrough] = await Promise.all([
     trpcClient.query('sudokus:get', { id: sudokuId }),
     trpcClient.query('walkthroughs:get', { sudokuId })
@@ -12,5 +15,5 @@ export const load: PageLoad = async ({ fetch, params }) => {
   if (sudoku == null) {
     throw error(404, 'Not found');
   }
-  return { sudoku, walkthrough };
+  return { sudoku, walkthrough, gameData: dataParam ? decompressFromBase64(dataParam.replace(/ /g, '+')) as SolutionStep : null };
 };
